# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Penguin Statistics v3 API, re-designed to aim for lightweight on wire.",
        "title": "Penguin Statistics API",
        "contact": {
            "name": "AlvISs_Reimu",
            "url": "https://github.com/AlvISsReimu",
            "email": "alvissreimu@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "3.0.0-alpha.1"
    },
    "host": "https://penguin-stats.io",
    "basePath": "/api",
    "paths": {
...
        "/PenguinStats/api/v2/report": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Submit an Item Drop Report",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/shims.ReportResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing itemId. Notice that this shall be the **string ID** of the item, instead of the internally used numerical ID of the item.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/PenguinStats/api/v2/report/recall": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Recall an Item Drop Report",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid or missing itemId. Notice that this shall be the **string ID** of the item, instead of the internally used numerical ID of the item.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/PenguinStats/api/v2/report/recognition": {
            "post": {
                "description": "Submit an Item Drop Report with Frontend Recognition. Notice that this is a private API and is not designed for external use.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Bulk Submit with Frontend Recognition",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Item"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "existence": {
                                            "$ref": "#/definitions/models.Existence"
                                        },
                                        "keywords": {
                                            "$ref": "#/definitions/models.Keywords"
                                        },
                                        "name": {
                                            "$ref": "#/definitions/models.I18nString"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing itemId. Notice that this shall be the **string ID** of the item, instead of the internally used numerical ID of the item.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
...
        "/PenguinStats/api/v2/stages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stage"
                ],
                "summary": "Get All Stages",
                "deprecated": true,
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/shims.Stage"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "code_i18n": {
                                                "$ref": "#/definitions/models.I18nString"
                                            },
                                            "existence": {
                                                "$ref": "#/definitions/models.Existence"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/PenguinStats/api/v2/stages/{stageId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stage"
                ],
                "summary": "Get an Stage with ID",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage ID",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/shims.Stage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code_i18n": {
                                            "$ref": "#/definitions/models.I18nString"
                                        },
                                        "existence": {
                                            "$ref": "#/definitions/models.Existence"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing stageId. Notice that this shall be the **string ID** of the stage, instead of the internally used numerical ID of the stage.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
...
        "/PenguinStats/api/v2/users": {
            "post": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Login with PenguinID",
                "deprecated": true,
                "parameters": [
                    {
                        "description": "User ID",
                        "name": "userId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/shims.SiteStats"
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
...
        "/v3/stages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stage"
                ],
                "summary": "Get All Stages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/models.Stage"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "code": {
                                                "$ref": "#/definitions/models.I18nString"
                                            },
                                            "existence": {
                                                "$ref": "#/definitions/models.Existence"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/v3/stages/{stageId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stage"
                ],
                "summary": "Get an Stage with ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage ID",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Stage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/models.I18nString"
                                        },
                                        "existence": {
                                            "$ref": "#/definitions/models.Existence"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing stageId. Notice that this shall be the **string ID** of the stage, instead of the internally used numerical ID of the stage.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.Extras": {
            "type": "object",
            "additionalProperties": true
        },
        "errors.PenguinError": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "extras": {
                    "$ref": "#/definitions/errors.Extras"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request: some or all request parameters are invalid"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "models.Existence": {
            "type": "object",
            "required": [
                "CN",
                "JP",
                "KR",
                "US"
            ],
            "properties": {
                "CN": {
                    "description": "CN: 国服 Mainland China Server (maintained by Hypergryph Network Technology Co., Ltd.)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "JP": {
                    "description": "JP: 日服 Japan Server (maintained by Yostar Inc,.)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "KR": {
                    "description": "KR: 韩服 Korea Server (maintained by Yostar Limited)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "US": {
                    "description": "US: 美服/国际服 Global Server (maintained by Yostar Limited)",
                    "$ref": "#/definitions/models.ServerExistence"
                }
            }
        },
        "models.I18nOptionalString": {
            "type": "object",
            "properties": {
                "en": {
                    "description": "EN: English (en)",
                    "type": "string"
                },
                "ja": {
                    "description": "JP: 日本語 (ja)",
                    "type": "string"
                },
                "ko": {
                    "description": "KR: 한국어 (ko)",
                    "type": "string"
                },
                "zh": {
                    "description": "ZH: 中文 (zh-CN)",
                    "type": "string"
                }
            }
        },
        "models.I18nString": {
            "type": "object",
            "required": [
                "en",
                "ja",
                "ko",
                "zh"
            ],
            "properties": {
                "en": {
                    "description": "EN: English (en)",
                    "type": "string"
                },
                "ja": {
                    "description": "JP: 日本語 (ja)",
                    "type": "string"
                },
                "ko": {
                    "description": "KR: 한국어 (ko)",
                    "type": "string"
                },
                "zh": {
                    "description": "ZH: 中文 (zh-CN)",
                    "type": "string"
                }
            }
        },
        "models.Item": {
            "type": "object",
            "properties": {
                "existence": {
                    "description": "Existence is a map with server code as key and the existence of the item in that server as value.",
                    "type": "object"
                },
                "group": {
                    "description": "Group is an identifier of what the item actually is. For example, both orirock and orirock cube would have the same group, `orirock`.",
                    "type": "string"
                },
                "itemId": {
                    "description": "ArkItemID (itemId) is the previously used, string form ID of the item; in JSON-representation `itemId` is used as key.",
                    "type": "string"
                },
                "keywords": {
                    "description": "Keywords is an arbitrary JSON object containing the keywords of the item, for optimizing the results of the frontend built-in search engine.",
                    "type": "object"
                },
                "name": {
                    "description": "Name is a map with language code as key and the name of the item in that language as value.",
                    "type": "object"
                },
                "penguinItemId": {
                    "description": "ItemID (penguinItemId) is the numerical ID of the item.",
                    "type": "integer"
                },
                "rarity": {
                    "type": "integer"
                },
                "sortId": {
                    "description": "SortID is the sort position of the item.",
                    "type": "integer"
                },
                "sprite": {
                    "description": "Sprite describes the location of the item's sprite on the sprite image, in a form of Y:X.",
                    "type": "string"
                }
            }
        },
        "models.Keywords": {
            "type": "object",
            "properties": {
                "alias": {
                    "description": "Alias of the item,",
                    "$ref": "#/definitions/models.I18nOptionalString"
                },
                "pron": {
                    "description": "Pronounciation hints of the item",
                    "$ref": "#/definitions/models.I18nOptionalString"
                }
            }
        },
...
        "models.ServerExistence": {
            "type": "object",
            "required": [
                "exist"
            ],
            "properties": {
                "endTime": {
                    "type": "integer"
                },
                "exist": {
                    "type": "boolean"
                },
                "startTime": {
                    "type": "integer"
                }
            }
        },
        "models.Stage": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is a map with language code as key and the code of the stage in that language as value.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "existence": {
                    "description": "Existence is a map with server code as key and the existence of the item in that server as value.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "extraProcessType": {
                    "description": "ExtraProcessType is the type of extra process that is used in the stage, e.g. \"GACHABOX\".",
                    "type": "string"
                },
                "minClearTime": {
                    "description": "MinClearTime is the minimum time (in milliseconds as a duration) it takes to clear the stage, referencing from prts.wiki",
                    "type": "number"
                },
                "penguinStageId": {
                    "description": "StageID (penguinStageId) is the numerical ID of the stage.",
                    "type": "integer"
                },
                "sanity": {
                    "description": "Sanity is the sanity requirement for a full clear of the stage.",
                    "type": "number"
                },
                "stageId": {
                    "description": "ArkStageID (stageId) is the previously used, string form ID of the stage; in JSON-representation `stageId` is used as key.",
                    "type": "string"
                },
                "stageType": {
                    "description": "StageType is the type of the stage, e.g. \"MAIN\", \"SUB\", \"ACTIVITY\" and \"DAILY\".",
                    "type": "string"
                },
                "zoneId": {
                    "description": "ZoneID is the numerical ID of the zone the stage is in.",
                    "type": "integer"
                }
            }
        },
...
        "shims.DropInfo": {
            "type": "object",
            "properties": {
                "bounds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "dropType": {
                    "type": "string"
                },
                "itemId": {
                    "type": "string"
                }
            }
        },
...
        "shims.ReportResponse": {
            "type": "object",
            "properties": {
                "reportHash": {
                    "type": "string"
                }
            }
        },
        "shims.SiteStats": {
            "type": "object",
            "properties": {
                "totalApCost": {
                    "type": "integer"
                },
                "totalItemQuantities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shims.TotalItemQuantity"
                    }
                },
                "totalStageTimes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shims.TotalStageTime"
                    }
                },
                "totalStageTimes_24h": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shims.TotalStageTime"
                    }
                }
            }
        },
        "shims.Stage": {
            "type": "object",
            "properties": {
                "apCost": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "code_i18n": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "dropInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shims.DropInfo"
                    }
                },
                "existence": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "minClearTime": {
                    "type": "number"
                },
                "recognitionOnly": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stageId": {
                    "type": "string"
                },
                "stageType": {
                    "type": "string"
                },
                "zoneId": {
                    "type": "string"
                }
            }
        },
...
        "shims.TotalItemQuantity": {
            "type": "object",
            "properties": {
                "itemId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "shims.TotalStageTime": {
            "type": "object",
            "properties": {
                "stageId": {
                    "type": "string"
                },
                "times": {
                    "type": "integer"
                }
            }
        },
...
    }
}